requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
motor>=3.3.2
alembic>=1.13.1
python-dateutil>=2.8.2
colorama>=0.4.6
rich>=13.7.0
transformers>=4.35.0
torch>=2.0.0
PyNaCl>=1.4.0
nltk>=3.8.1
aiofiles>=0.8.0
pymysql>=1.0.3
typing-extensions>=4.8.0
aiohttp>=3.9.0
textblob>=0.17.1
psutil>=5.9.0
python-magic>=0.4.27
beautifulsoup4>=4.12.0
markdown>=3.5.0
Pillow>=10.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
python-pdf>=0.39
emoji>=2.8.0
pytz>=2023.3
requests>=2.31.0
pywin32>=224